+++ src/Core/Dom/DomXpath.php
@@ -26,8 +26,21 @@ class DomXpath extends BaseDomXpath
         parent::__construct($doc->getDom());
     }
 
+    /**
+     * @inheritdoc
+     */
     public function query($expr, BaseDomNode $context = null, $registerNodeNS = null)
     {
+
+        // NullDomNode is an addition to SERPS and must be handled differently
+        if ($context instanceof NullDomNode) {
+            if ($expr{0} == '/') {
+                $context = null;
+            } else {
+                return new EmptyDomNodeList();
+            }
+        }
+
         $nodeList = parent::query($expr, $context);
         return new DomNodeList($nodeList, $this->documentWrapper);
     }